+++ UnitTests/ChallengeTests/Challenge1Test.cs
@@ -13,7 +13,7 @@ namespace Challenges
             Challenge1 challenge1 = new Challenge1();
             challenge1.CalculateUntilNotIncluding = 10;
 
-            int result = challenge1.RunChallenge();
+            long result = challenge1.RunChallenge();
 
             Assert.AreEqual(23, result);
         }
@@ -25,7 +25,7 @@ namespace Challenges
             Challenge1 challenge1 = new Challenge1();
             challenge1.CalculateUntilNotIncluding = 1000;
 
-            int result = challenge1.RunChallenge();
+            long result = challenge1.RunChallenge();
 
             Assert.AreEqual(233168, result);
         }
