+++ UnitTests/EnumeratorTests/NaturalTest.cs
@@ -17,7 +17,7 @@ namespace ProjectEuler
         {
             NaturalNumbers natNum = new NaturalNumbers();
 
-            int the123rdnumber = natNum.Skip(122).First();
+            long the123rdnumber = natNum.Skip(122).First();
             
             Assert.AreEqual(123, the123rdnumber);
